+++ scripts/reboot.sh
@@ -1,0 +1,4 @@
+#!/bin/bash -eu
+
+echo -e '\033[33mRebooting VM before running Integration Tests ...\033[0m'
+reboot